+++ angular_compiler/CHANGELOG.md
@@ -1,3 +1,7 @@
+* Removed all remaining (invalid) references to `package:barback`.
+
+## 0.4.0-alpha+5
+
 ### Bug fixes
 
 * `linkTypeOf` correctly resolves bound types (i.e. `<T>`) in most cases, and